+++ crates/repl/src/notebook.rs
@@ -1,4 +1,5 @@
 mod cell;
+mod env_name_modal;
 mod notebook_ui;
 pub use cell::*;
 pub use notebook_ui::*;